Pulled from upstream and modified for have login working. Changes in the settings for working with environment variables. Changed annotator_id column name in accounts/models.py (issue #57)
