+++ accounts/forms/__init__.py
@@ -1,3 +1,2 @@
 from register import RegistrationForm
 from authenticate import AuthenticationForm
-from profile import ProfileForm
