+++ b2note_devel/urls.py
@@ -10,8 +10,6 @@ urlpatterns = patterns('',
     url(r'^accounts/', include('accounts.urls', namespace='accounts')),
     url(r'^$', 'accounts.views.login', name='index'),
     url(r'^login', 'accounts.views.login'),
-
-    url(r'^profilepage$', 'b2note_app.views.profilepage', name='profile'),
     url(r'^hostpage', 'b2note_app.views.hostpage'),
     url(r'^interface_main', 'b2note_app.views.interface_main'),
     url(r'^create_annotation', 'b2note_app.views.create_annotation'),
